Fix GH-14067: Check for ZEND_MOD_CONFLICTS in zend_startup_module_ex
When extension has conflicting extension(s) listed in ZEND_MOD_CONFLICTS
this now properly emits a warning and doesn't load the extension.

Previously it relied on the sort order of the registered extensions:
- first the statically built extensions based on the
  PHP_ADD_EXTENSION_DEP (*nix) or ADD_EXTENSION_DEP (Windows)
  listed in the main/internal_functions.c or
  main/internal_functions_cli.c.
- then dynamically loaded extensions using the extension=<ext> and
  zend_extension=<ext> INI directives.

And it didn't detect conflicting extensions when they were loaded later
in the list or when loaded as shared modules.