ext/intl: don't always enforce c++17
the intention of this condition was to check if build is done against icu >= 74
while in practice it's a condition in the form of `test STRING` which in turn is
equivalent to `test -n STRING` that checks if passed string is not empty.
condition is always satisfied and hence c++17 is also always required.

strip `test` part and invoke `pkg-config` directly in condition.