Fix the aligned variable attribute check
By default compilers may not treat attribute warnings as errors when
encountering an unknown __attribute__, unless some error option is
provided (-Werror=attributes, -Werror...). This fixes the check and
wraps it into a separate M4 macro to be extendable in the future if
needed. It checks if conftest.err file was generated by the compilation
check when warnings appear. Also, PHP check is a bit customized by using
__alignof__ keyword, so it is left in there for now to not break
existing checks.