Fix reentrant functions declarations checks
This fixes incompatible pointer type warnings during the reentrant
functions declaration checks. These were not declared on some obsolete
systems if _REENTRANT was not defined. The check is for now left in the
code base but can be transitioned to newer code without checking for
missing declarations or using these otherwise in the future.