Squeeze PHP_CHECK_LIBRARY to reduce generated configure
This trims redundant newlines in the generated configure script.

The `:` is a shell builtin command that does nothing [1] used in
AC_CHECK_LIB to prevent doing its default action when library is found
and 3rd argument is not given. Supported by all today's shells where
Autoconf's configure works.

The 5th argument (ac_stuff) is left quoted with double quotes to
properly pass it further to PHP_EVAL_LIBLINE which has the for loop at
this point when given argument includes newlines.

[1] https://www.gnu.org/savannah-checkouts/gnu/bash/manual/bash.html#Bourne-Shell-Builtins