Revert "[ci skip] Add Hamada-san to CODEOWNERS for mbstring"
This reverts commit 0bcc1e613bc8b62859fc75afeeb83651555fb44a.

It turns out that GitHub expects all users listed in CODEOWNERS to have
write access to the repository, and @youkidearitai does not currently
have such access. I was not aware of this earlier, and still hope that
he can be added back later, since he is very helpful when it comes
to reviewing GitHub issues and pull requests.