Move ext/tidy to pkg-config
It seems most distributions i.e. Debian/Gentoo/Fedora have all moved
over to html-tidy as the tidy library they ship. It ships a pkg-config
file, so we can just simply check for that and omit the various
alternatives to check for.

I'm not sure if the checks for library functionality are needed anymore
if we assume the html5 tidy fork either.